allow full package names for Plack Middleware
`Plack::Builder` only prefixes the provided package name with
`Plack::Middleware::` if it does not already have that as a prefix.

Thus, `enable 'Plack::Middleware::Foo'` was parsing here and adding
_another_ `P:M:` to the start, giving us `P:M:P:M:Foo`.

Easiest fix is to strip the prefix if it already exists, as we're going
to add it back on again in the next line anyways.